refactor: 重写混入类
